requirements.txt
Flask
Flask-Cors
Pillow
transformers
gunicorn
supabase
timm
requests
opencv-python-headless
--extra-index-url https://download.pytorch.org/whl/cpu
torch